new data, improved upload preparation, file structure changed
scraper and converter scripts now organise files in folders based on
dates, upload preparation script will only prepare output for a given
date for upload

+++ media/eu_parliament_audiovisual/combine_ntriples2ntriples.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
-rapper -i turtle handcrafted_rdf/known_topics.n3 > output/known_topics.nt
-rapper -i turtle handcrafted_rdf/other.ttl > output/other.nt
-cat output/*/*.nt output/*.nt > eup_media_assets.nt
-split -l 9000 eup_media_assets.nt eup_media.
-rm upload/*
-mv eup_media.* upload
+if [ "$#" -gt 0 ]; then
+	rapper -i turtle handcrafted_rdf/known_topics.n3 > output/known_topics.nt
+	rapper -i turtle handcrafted_rdf/other.ttl > output/other.nt
+	cat output/$1/*.nt output/*.nt > eup_media_assets.nt
+	split -l 9000 eup_media_assets.nt eup_media.
+	rm upload/*
+	mv eup_media.* upload
+else
+	echo "You need to specify a date!"
+fi
+
